# docs/swagger.yaml
basePath: /
definitions:
  domain.Message:
    properties:
      id:
        type: integer
      message:
        type: string
      room_id:
        type: string
      timestamp:
        type: string
      user_id:
        type: integer
    type: object
  domain.Room:
    properties:
      created_at:
        type: string
      id:
        type: string
      room_name:
        type: string
    type: object
  http.CreateRoomRequest:
    properties:
      room_name:
        type: string
    type: object
  http.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  http.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: goldenglowitsolutions@gmail.com
    name: API Support
    url: https://www.josuebarros.com/support
  description: This is a Golang Chat API for real-time chat.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Golang Chat API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login a user
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - users
  /rooms:
    get:
      description: Retrieve all available rooms for users to join
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Room'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a list of available chat rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Create a room for users to join
      parameters:
      - description: Room Info
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/http.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new chat room
      tags:
      - rooms
  /rooms/{roomID}/messages:
    get:
      description: Fetch the last 50 messages from a specified room
      parameters:
      - description: Room ID
        in: path
        name: roomID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Message'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get messages from a specific chat room
      tags:
      - messages
  /ws/{roomID}:
    get:
      description: Connect to a WebSocket for real-time communication in a room
      parameters:
      - description: Room ID
        in: path
        name: roomID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: WebSocket Connection Established
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Establish a WebSocket connection
      tags:
      - websocket
swagger: "2.0"